+++ R/tidyquant.R
@@ -28,13 +28,12 @@
 #' @name tidyquant
 #'
 #' @import quantmod
-#' @import scales
 #' @import tidyverse
 #' @import lubridate
 #' @import PerformanceAnalytics
+#' @importFrom rlang ":="
 #' @importFrom magrittr "%>%"
 #' @importFrom magrittr "%$%"
-#' @importFrom stats "setNames"
 #' @importFrom utils "data" "download.file" "read.csv"
 #' @importFrom TTR "SMA"
 #' @importFrom xts "to.period"